requirements.txt
asgiref==3.5.1
backports.zoneinfo==0.2.1
bcrypt==3.2.2
certifi==2021.10.8
charset-normalizer==2.0.12
cffi==1.15.0
cryptography==37.0.2
Django==4.0.4
django-cors-headers==3.11.0
idna==3.3
mysqlclient==2.1.0
pycparser==2.21
PyJWT==2.4.0
PyMySQL==1.0.2
requests==2.27.1
sqlparse==0.4.2
urllib3==1.26.9
